Add backpressure support
The `give`, `giveBack`, `error` and `close` methods now
return promises to allow the producer side of the channel
to wait until more data is requested by the consumer side.

The use of these promises is not required. It only exists
to serve as an optimization helper to prevent unbounded
memory growth in the event the producer outruns to consumer.